requirements.txt
google-adk==1.19.0
google-auth
pandas
streamlit
tabulate
